pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "second-order-by-annealer"
version = "0.1.0"
description = "Quadratic annealing optimizer experiments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "scikit-learn>=1.2",
    "dimod>=0.12",
    "dwave-system>=1.10",
    "dwave-samplers>=0.3",
    "tqdm>=4.65",
]

[tool.setuptools.packages.find]
include = ["src*"]